Aggiunto promozione con finestra modal
